==> code/user_newt/CIKM-2019/CIKM-2019-AnalytiCup-有注释/Semi-Finals/underline_trainning/Step1 itemCF_based_on_Apriori/2_generate_hot_table.ipynb <==
## 生成中间文件

本代码的目的是生成三个中间文件，其中hot_map是统计的商品的出现次数，upwardmap与downward_map是将商品id映射到实数集\[0, m\]，其中m代表商品总数。

This code aims to generate three temporary files. Hot_map statistics the number of appearance of each item. Upward_map and Downward_map can map the ItemID into \[0, m\], where m indicates the size of item corpus.

import pandas as pd
import numpy as np
import os
# round2 train的路径
path =r'C:\Users\NewtScamander\github\CIKM-2019-AnalytiCup\data\ECommAI_EUIR_round1_testA_20190701'

data = pd.read_csv(os.path.join(path, 'user_behavior.csv'),header=None)
item=pd.read_csv(os.path.join(path, 'item.csv'), header=None)
data.columns = ['userID','itemID','behavior','timestamp']
item.columns = ['itemID', 'category', 'shop', 'band']

data['day'] = data['timestamp'] // 86400
data['hour'] = data['timestamp'] // 3600 % 24

user_times = data[['itemID','userID']].groupby('userID', as_index=False).count()

user_times.columns = ['userID','itemCount']

user_times.head()

user_times_map = dict(zip(user_times['userID'], user_times['itemCount'])) # 用户活跃度 {'userID': 'itemCount'}

len(user_times_map)

f = open('usersActivity_map.txt', 'w')
f.write(str(user_times_map))
f.close()

生成upward_map 与 downward_map:

from sklearn import preprocessing
# 对itemID进行编码，比如[2311,2312,2313]->[1,2,3]
le = preprocessing.LabelEncoder()
item['encoding'] = le.fit_transform(item['itemID'])
upward_map = dict(zip(item['itemID'], item['encoding']))
downward_map = dict(zip(item['encoding'], item['itemID']))

生成hot table

temp = data[['itemID','behavior']].groupby('itemID',as_index=False).count()
hot_map = dict(zip(temp['itemID'], temp['behavior']))

hot_map

def save_to_file(trans_map, file_path):
    trans_map = str(trans_map)
    f = open(file_path, 'w')
    f.write(trans_map)
    f.close()

save_to_file(hot_map,'hot_items_map.txt')

save_to_file(upward_map,'upward_map.txt')

save_to_file(downward_map,'downward_map.txt')
