==> code/user_newt/CIKM-2019/CIKM-2019-AnalytiCup-有注释/Semi-Finals/underline_trainning/Step1 itemCF_based_on_Apriori/3_generate_original_matrix_0.ipynb <==
## 统计相似度矩阵

此代码目的是统计每个group中的相似度矩阵，由于cython和multiprocessing在notebook环境下同时使用会出现一些问题，所以分了很多个文件。

This code aims to calculate the similarity matrix in each group. Due to cython and multiprocessing cannot work well together in notebook envirment, so we generated many simliar codes to finish the same work for multiprocessing.

import pandas as pd
import numpy as np
import sys
from scipy.sparse import lil_matrix
import scipy as scp
import time
%load_ext Cython

ITEM_NUM = 4318201

def get_logs_from_hardisk(path):
    f = open(path, 'r')
    a = f.read()
    dict_name = eval(a)
    f.close()
    return dict_name


f = open('usersActivity_map.txt', 'r')
m = f.read()
user_times_map = eval(m)  # user_times_map 用户活跃度字典{'userID':'itemCount'}
f.close()

%%cython

import datetime
import math

cpdef calculate_matrix(mat, list user_logs, dict user_times_map):
    # mat物品相似度的稀疏矩阵， user_logs用户对应商品列表的字典{'userID':[itemID1,itemID2,...]},user_times_map用户活跃度字典
    cdef int index, i1, i2, count
    cdef list item_log
    cdef tuple u
    
    count = 0
    for u in user_logs:
        count += 1
        if count % 1000 == 0:
            print('The %d'%count + ' users are finished.')
            print(datetime.datetime.now().strftime('%H:%M:%S'))
            
        item_log = u[1]   

        for index, i1 in enumerate(item_log):
            for i2 in item_log[(index+1): ]:
                weight = 1/(math.log(1+user_times_map[u[0]])) # Inverse User Frequence 用户活跃度越大对物品相似度贡献越小
                mat[i1, i2] += weight
                mat[i2, i1] += weight
    return mat

f = open('upward_map.txt','r')
upward_map = eval(f.read())  # upward_map商品ID到商品编码的字典
f.close()

user_logs = get_logs_from_hardisk('userlogs_group0.txt')
for u in user_logs:  
    user_logs[u] = [int(upward_map[x]) for x in user_logs[u]] #商品ID映射到商品编码
    
user_logs = list(user_logs.items())
print(user_logs[0])

for i in range(0, len(user_logs), 10000):
    print('The %d '%i + ' batch is started!')
    print('--------------------------------')
    mat = lil_matrix((ITEM_NUM+1, ITEM_NUM+1), dtype=float)
    mat = calculate_matrix(mat, user_logs[i: i + 10000], user_times_map)
    scp.sparse.save_npz('tmpData\\sparse_matrix_%d_batch_group0.npz'%i, mat.tocsr())
    print('save successfully')
    print('--------------------------------')
